requirements.txt
fastapi uvicorn azure-storage-blob python-multipart
python-dotenv